refactor: move DRTPLayer import from core to hooks in decoders.py and remove unused imports in __init__.py
