+++ src/ehc_sn/core/__init__.py
@@ -5,7 +5,3 @@ This module contains base classes for building and training neural network model
 for spatial navigation, including components for both hippocampal and entorhinal cortex
 representations.
 """
-
-from .drtp import DRTPFunction, DRTPLayer, drtp_layer
-
-__all__ = ["DRTPFunction", "DRTPLayer", "drtp_layer"]
